+++ CHANGELOG.md
@@ -8,6 +8,11 @@
   - <https://github.com/vivliostyle/vivliostyle.js/pull/323>
   - Spec proposal: [CSS Repeated Headers and Footers](https://specs.rivoal.net/css-repeat/)
 
+### Changed
+
+- Place footnotes at the bottom of pages (or regions) rather than columns
+  - <https://github.com/vivliostyle/vivliostyle.js/pull/343>
+
 ### Fixed
 
 - Fix footnote layout bugs